[SlashSync] Cog model restructure
- Restructure `SlashSync` according to SFUAnime/Ren#608. No functional
  changes.
- Remove module docstring for `slashsync.py` as it is already included
  in `__init__.py`.
